iot-client/requirements.txt
azure-iot-device==2.12.0
python-dotenv==1.0.0
azure-eventhub==5.10.1
flask
flask_cors
